modified multiply solution

+++ multiply.c
@@ -1,9 +1,9 @@
-#include<stdio.h>
+#include <stdio.h>
 int main()
 {
-    long int A,B;
-    scanf("%ld %ld",&A,&B);
-    long int multiplication = A*B;
-    printf("%ld",multiplication);
-    return 0;    
+    long long int A, B;
+    scanf("%lld %lld", &A, &B);
+    long long int multiplication = A * B;
+    printf("%lld", multiplication);
+    return 0;
 }
